+++ libs/cmcd/README.md
@@ -24,23 +24,23 @@ const result = encodeCmcd(input, options);
 // result === 'br=1000,com.example-hello="world",ec=("ERR001" "ERR002"),su,v=2'
 ```
 
-## Testing CMCD output with `CmcdRequestCollector`
+## Testing CMCD output with `CmcdReportRecorder`
 
-`CmcdRequestCollector` captures CMCD-bearing requests emitted by a
+`CmcdReportRecorder` captures CMCD-bearing requests emitted by a
 player under test, regardless of whether the player uses
 `XMLHttpRequest` or `fetch`. Each capture is normalized to
 [`HttpRequest`](https://github.com/streaming-video-technology-alliance/common-media-library/tree/main/libs/utils),
 so tests work identically across transports.
 
 ````ts
-import { CmcdRequestCollector, CmcdRequestType, validateCmcdRequest } from '@svta/cml-cmcd'
+import { CmcdReportRecorder, validateCmcdRequest } from '@svta/cml-cmcd'
 
-const collector = new CmcdRequestCollector()
-collector.attach({ eventTargetUrls: ['https://events.example.com'] })
+const recorder = new CmcdReportRecorder()
+recorder.attach({ eventTargetUrls: ['https://events.example.com'] })
 
 // ... configure and start the player under test ...
 
-const segments = await collector.waitForRequests(CmcdRequestType.SEGMENT, 3)
+const segments = await recorder.waitForSegments({ count: 3 })
 for (const r of segments) {
   const result = validateCmcdRequest(r.request)
   if (!result.valid) {
@@ -48,8 +48,5 @@ for (const r of segments) {
   }
 }
 
-collector.detach()
+recorder.detach()
 ````
-
-Use `collectFor(timeout, type?)` instead of `waitForRequests` when you
-need to verify that requests did *not* arrive within a window.
