+++ libs/cmcd/config/cml-cmcd.api.md
@@ -23,15 +23,6 @@ export function appendCmcdQuery(url: string, cmcd: Cmcd, options?: CmcdEncodeOpt
 export type Cmcd = CmcdRequest & CmcdResponse & CmcdEvent;
 
 // @public
-export const CMCD_COLLECTED_REQUEST_MODE_EVENT: "event";
-
-// @public
-export const CMCD_COLLECTED_REQUEST_MODE_HEADER: "header";
-
-// @public
-export const CMCD_COLLECTED_REQUEST_MODE_QUERY: "query";
-
-// @public
 export const CMCD_DEFAULT_TIME_INTERVAL = 30;
 
 // @public
@@ -122,6 +113,27 @@ export const CMCD_PARAM = "CMCD";
 export const CMCD_QUERY: "query";
 
 // @public
+export const CMCD_RECORDED_REPORT_MODE_EVENT: "event";
+
+// @public
+export const CMCD_RECORDED_REPORT_MODE_HEADER: "header";
+
+// @public
+export const CMCD_RECORDED_REPORT_MODE_QUERY: "query";
+
+// @public
+export const CMCD_RECORDER_REQUEST_TYPE_EVENT: "event";
+
+// @public
+export const CMCD_RECORDER_REQUEST_TYPE_MANIFEST: "manifest";
+
+// @public
+export const CMCD_RECORDER_REQUEST_TYPE_SEGMENT: "segment";
+
+// @public
+export const CMCD_RECORDER_REQUEST_TYPE_UNKNOWN: "unknown";
+
+// @public
 export const CMCD_REQUEST: "CMCD-Request";
 
 // @public
@@ -129,18 +141,6 @@ export const CMCD_REQUEST_KEYS: readonly ["ab", "bg", "bl", "br", "bs", "bsa", "
 
 // @public
 export const CMCD_REQUEST_MODE: "request";
-
-// @public
-export const CMCD_REQUEST_TYPE_EVENT: "event";
-
-// @public
-export const CMCD_REQUEST_TYPE_MANIFEST: "manifest";
-
-// @public
-export const CMCD_REQUEST_TYPE_SEGMENT: "segment";
-
-// @public
-export const CMCD_REQUEST_TYPE_UNKNOWN: "unknown";
 
 // @public
 export const CMCD_RESPONSE_KEYS: readonly ["cmsdd", "cmsds", "rc", "smrt", "ttfb", "ttfbb", "ttlb", "url"];
@@ -165,24 +165,6 @@ export const CMCD_VALIDATION_SEVERITY_ERROR: "error";
 
 // @public
 export const CMCD_VALIDATION_SEVERITY_WARNING: "warning";
-
-// @public
-export type CmcdCollectedRequest = {
-    readonly request: HttpRequest;
-    readonly type: CmcdRequestType;
-    readonly reportingMode: CmcdCollectedRequestMode;
-    readonly timestamp: number;
-};
-
-// @public
-export const CmcdCollectedRequestMode: {
-    readonly QUERY: typeof CMCD_COLLECTED_REQUEST_MODE_QUERY;
-    readonly HEADER: typeof CMCD_COLLECTED_REQUEST_MODE_HEADER;
-    readonly EVENT: typeof CMCD_COLLECTED_REQUEST_MODE_EVENT;
-};
-
-// @public (undocumented)
-export type CmcdCollectedRequestMode = ValueOf<typeof CmcdCollectedRequestMode>;
 
 // @public
 export type CmcdCustomKey = `${string}-${string}`;
@@ -333,6 +315,35 @@ export const CmcdPlayerState: {
 export type CmcdPlayerState = ValueOf<typeof CmcdPlayerState>;
 
 // @public
+export type CmcdRecordedReport = {
+    readonly request: HttpRequest;
+    readonly type: CmcdRecorderRequestType;
+    readonly reportingMode: CmcdRecordedReportMode;
+    readonly timestamp: number;
+};
+
+// @public
+export const CmcdRecordedReportMode: {
+    readonly QUERY: typeof CMCD_RECORDED_REPORT_MODE_QUERY;
+    readonly HEADER: typeof CMCD_RECORDED_REPORT_MODE_HEADER;
+    readonly EVENT: typeof CMCD_RECORDED_REPORT_MODE_EVENT;
+};
+
+// @public (undocumented)
+export type CmcdRecordedReportMode = ValueOf<typeof CmcdRecordedReportMode>;
+
+// @public
+export const CmcdRecorderRequestType: {
+    readonly MANIFEST: typeof CMCD_RECORDER_REQUEST_TYPE_MANIFEST;
+    readonly SEGMENT: typeof CMCD_RECORDER_REQUEST_TYPE_SEGMENT;
+    readonly EVENT: typeof CMCD_RECORDER_REQUEST_TYPE_EVENT;
+    readonly UNKNOWN: typeof CMCD_RECORDER_REQUEST_TYPE_UNKNOWN;
+};
+
+// @public (undocumented)
+export type CmcdRecorderRequestType = ValueOf<typeof CmcdRecorderRequestType>;
+
+// @public
 export type CmcdReportConfig = {
     version?: CmcdVersion;
     enabledKeys?: CmcdKey[];
@@ -372,6 +383,32 @@ export const CmcdReportingMode: {
 
 // @public (undocumented)
 export type CmcdReportingMode = ValueOf<typeof CmcdReportingMode>;
+
+// @public
+export class CmcdReportRecorder {
+    attach(options?: CmcdReportRecorderOptions): void;
+    clear(): void;
+    detach(): void;
+    getReports(): CmcdRecordedReport[];
+    waitForEvents(options?: CmcdReportRecorderWaitOptions): Promise<CmcdRecordedReport[]>;
+    waitForManifest(options?: CmcdReportRecorderWaitOptions): Promise<CmcdRecordedReport[]>;
+    waitForReports(options?: CmcdReportRecorderWaitOptions): Promise<CmcdRecordedReport[]>;
+    waitForSegments(options?: CmcdReportRecorderWaitOptions): Promise<CmcdRecordedReport[]>;
+}
+
+// @public
+export type CmcdReportRecorderOptions = {
+    eventTargetUrls?: readonly string[];
+    transports?: readonly CmcdTransportAdapter[];
+    waitTimeout?: number;
+    onReport?: (report: CmcdRecordedReport) => void;
+};
+
+// @public
+export type CmcdReportRecorderWaitOptions = {
+    count?: number;
+    timeout?: number;
+};
 
 // @public
 export type CmcdRequest = {
@@ -419,22 +456,6 @@ export type CmcdRequest = {
 };
 
 // @public
-export class CmcdRequestCollector {
-    attach(options?: CmcdRequestCollectorOptions): void;
-    clear(): void;
-    collectFor(timeout: number, type?: CmcdRequestType): Promise<CmcdCollectedRequest[]>;
-    detach(): void;
-    getRequests(type?: CmcdRequestType): CmcdCollectedRequest[];
-    waitForRequests(type: CmcdRequestType | undefined, count: number, timeout?: number): Promise<CmcdCollectedRequest[]>;
-}
-
-// @public
-export type CmcdRequestCollectorOptions = {
-    eventTargetUrls?: readonly string[];
-    transports?: readonly CmcdTransportAdapter[];
-};
-
-// @public
 export type CmcdRequestDeliver = (request: HttpRequest) => Response | undefined;
 
 // @public
@@ -452,17 +473,6 @@ export type CmcdRequestReport<D = unknown> = HttpRequest & {
 export type CmcdRequestReportConfig = CmcdReportConfig & {
     transmissionMode?: CmcdTransmissionMode;
 };
-
-// @public
-export const CmcdRequestType: {
-    readonly MANIFEST: typeof CMCD_REQUEST_TYPE_MANIFEST;
-    readonly SEGMENT: typeof CMCD_REQUEST_TYPE_SEGMENT;
-    readonly EVENT: typeof CMCD_REQUEST_TYPE_EVENT;
-    readonly UNKNOWN: typeof CMCD_REQUEST_TYPE_UNKNOWN;
-};
-
-// @public (undocumented)
-export type CmcdRequestType = ValueOf<typeof CmcdRequestType>;
 
 // @public
 export type CmcdResponse = CmcdRequest & {
