+++ libs/cmcd/src/index.ts
@@ -18,8 +18,6 @@ export * from './CMCD_RESPONSE_KEYS.ts'
 export * from './CMCD_V1.ts'
 export * from './CMCD_V1_KEYS.ts'
 export * from './CMCD_V2.ts'
-export type * from './CmcdCollectedRequest.ts'
-export * from './CmcdCollectedRequestMode.ts'
 export type * from './CmcdCustomKey.ts'
 export type * from './CmcdCustomValue.ts'
 export type * from './CmcdData.ts'
@@ -41,17 +39,20 @@ export type * from './CmcdKey.ts'
 export * from './CmcdObjectType.ts'
 export type * from './CmcdObjectTypeList.ts'
 export * from './CmcdPlayerState.ts'
+export type * from './CmcdRecordedReport.ts'
+export * from './CmcdRecordedReportMode.ts'
 export type * from './CmcdReportConfig.ts'
+export * from './CmcdReportRecorder.ts'
+export type * from './CmcdReportRecorderOptions.ts'
+export type * from './CmcdReportRecorderWaitOptions.ts'
 export * from './CmcdReporter.ts'
 export type * from './CmcdReporterConfig.ts'
 export * from './CmcdReportingMode.ts'
 export type * from './CmcdRequest.ts'
-export * from './CmcdRequestCollector.ts'
-export type * from './CmcdRequestCollectorOptions.ts'
 export type * from './CmcdRequestKey.ts'
 export type * from './CmcdRequestReport.ts'
 export type * from './CmcdRequestReportConfig.ts'
-export * from './CmcdRequestType.ts'
+export * from './CmcdRecorderRequestType.ts'
 export type * from './CmcdResponse.ts'
 export * from './CmcdStreamingFormat.ts'
 export * from './CmcdStreamType.ts'
